@@ -2,6 +2,7 @@
   "name": "@bigcommerce/eslint",
   "version": "1.0.1",
   "description": "Default ESLint configuration used at BigCommerce",
+  "type": "module",
   "private": true,
   "author": "BigCommerce",
   "license": "MIT",
@@ -16,14 +17,16 @@
   "scripts": {
     "lint": "eslint . --ext .js",
     "release": "lerna publish --sign-git-commit --sign-git-tag --git-remote origin --pre-dist-tag next",
-    "test": "jest --projects packages/*"
+    "test": "NODE_OPTIONS='--experimental-vm-modules' jest --projects packages/*"
   },
   "devDependencies": {
-    "eslint": "^8.14.0",
+    "eslint": "^9.0.0",
     "jest": "^29.7.0",
     "lerna": "^8.1.2",
     "typescript": "^5.0.2"
   },
-  "dependencies": {},
+  "dependencies": {
+    "globals": "^15.0.0"
+  },
   "packageManager": "yarn@1.22.22+sha256.c17d3797fb9a9115bf375e31bfd30058cac6bc9c3b8807a3d8cb2094794b51ca"
 }
