@@ -2,11 +2,11 @@
   "name": "@bigcommerce/eslint-config",
   "version": "2.11.0",
   "description": "Default ESLint configuration used at BigCommerce",
+  "type": "module",
   "main": "index.js",
   "files": [
     "index.js",
     "configs",
-    "patch",
     "plugin",
     "utils",
     "prettier",
@@ -21,7 +21,7 @@
   },
   "dependencies": {
     "@bigcommerce/eslint-plugin": "^1.4.0",
-    "@rushstack/eslint-patch": "^1.10.5",
+    "@eslint/js": "^9.0.0",
     "@stylistic/eslint-plugin": "^2.1.0",
     "@typescript-eslint/eslint-plugin": "^8.26.0",
     "@typescript-eslint/parser": "^8.26.0",
@@ -39,6 +39,7 @@
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-switch-case": "^1.1.2",
     "eslint-plugin-testing-library": "^7.1.1",
+    "globals": "^15.0.0",
     "prettier": "^3.5.3"
   },
   "devDependencies": {
@@ -47,7 +48,7 @@
     "react": "^18.3.1"
   },
   "peerDependencies": {
-    "eslint": "^8.0.0",
+    "eslint": "^9.0.0",
     "typescript": "^4.0.0 || ^5.0.0"
   },
   "peerDependenciesMeta": {
