@@ -16,9 +16,10 @@
   "dependencies": {
     "tsutils": "^3.21.0"
   },
+  "type": "module",
   "peerDependencies": {
     "@typescript-eslint/parser": "^8.0.0",
-    "eslint": "^8.0.0",
+    "eslint": "^9.0.0",
     "typescript": "^4.0.0 || ^5.0.0"
   },
   "devDependencies": {
