@@ -1,336 +1,352 @@
-/** @type {import("eslint").Linter.Config} */
-module.exports = {
-  env: {
-    browser: true,
-    es6: true,
+import js from '@eslint/js';
+import stylisticPlugin from '@stylistic/eslint-plugin';
+import gettextPlugin from 'eslint-plugin-gettext';
+import importPlugin from 'eslint-plugin-import';
+import switchCasePlugin from 'eslint-plugin-switch-case';
+import globals from 'globals';
+
+export default [
+  js.configs.recommended,
+  {
+    languageOptions: {
+      ecmaVersion: 'latest',
+      sourceType: 'module',
+      globals: {
+        ...globals.browser,
+        ...globals.es2021,
+      },
+    },
+    plugins: {
+      import: importPlugin,
+      gettext: gettextPlugin,
+      'switch-case': switchCasePlugin,
+      '@stylistic': stylisticPlugin,
+    },
+    settings: {},
+    linterOptions: {
+      reportUnusedDisableDirectives: 'error',
+    },
+    rules: {
+      '@stylistic/padding-line-between-statements': [
+        'warn',
+        {
+          blankLine: 'always',
+          next: 'return',
+          prev: '*',
+        },
+        {
+          blankLine: 'always',
+          next: ['block', 'block-like', 'class', 'const', 'let', 'var', 'interface'],
+          prev: '*',
+        },
+        {
+          blankLine: 'always',
+          next: '*',
+          prev: ['block', 'block-like', 'class', 'const', 'let', 'var', 'interface'],
+        },
+        {
+          blankLine: 'any',
+          next: ['const', 'let', 'var'],
+          prev: ['const', 'let', 'var'],
+        },
+        {
+          blankLine: 'any',
+          next: ['case'],
+          prev: ['case'],
+        },
+        {
+          blankLine: 'always',
+          next: '*',
+          prev: 'directive',
+        },
+        {
+          blankLine: 'any',
+          next: 'directive',
+          prev: 'directive',
+        },
+      ],
+      '@stylistic/spaced-comment': [
+        'warn',
+        'always',
+        {
+          block: { balanced: true, exceptions: ['-', '+'], markers: ['=', '!'] },
+          line: { exceptions: ['-', '+'], markers: ['=', '!'] },
+        },
+      ],
+      'array-callback-return': 'error',
+      'arrow-body-style': ['error', 'as-needed', { requireReturnForObjectLiteral: false }],
+      'block-scoped-var': 'error',
+      camelcase: ['error', { properties: 'never' }],
+      complexity: 'error',
+      'consistent-return': 'error',
+      'default-case': ['error'],
+      'default-param-last': 'error',
+      'dot-notation': ['error', { allowKeywords: true }],
+      eqeqeq: ['error', 'smart'],
+      'func-names': 'error',
+      'getter-return': ['error', { allowImplicit: true }],
+      'gettext/no-variable-string': 'error',
+      'guard-for-in': 'error',
+      'import/default': 'off',
+      'import/dynamic-import-chunkname': 'error',
+      'import/extensions': [
+        'error',
+        {
+          js: 'never',
+          json: 'always',
+          jsx: 'never',
+          ts: 'never',
+          tsx: 'never',
+        },
+      ],
+      'import/newline-after-import': 'warn',
+      'import/no-absolute-path': 'error',
+      'import/no-amd': 'error',
+      'import/no-duplicates': ['warn', { 'prefer-inline': true }],
+      'import/no-dynamic-require': 'error',
+      'import/no-extraneous-dependencies': 'error',
+      'import/no-mutable-exports': 'error',
+      'import/no-named-default': 'error',
+      'import/no-unresolved': ['error', { caseSensitive: true, commonjs: true }],
+      'import/no-webpack-loader-syntax': 'error',
+      'import/order': [
+        'warn',
+        {
+          alphabetize: { caseInsensitive: true, order: 'asc' },
+          groups: [['builtin', 'external'], 'internal', 'parent', 'sibling', 'index'],
+          'newlines-between': 'always',
+        },
+      ],
+      'max-classes-per-file': ['error', 1],
+      'new-cap': [
+        'error',
+        {
+          capIsNew: false,
+          capIsNewExceptions: ['Immutable.Map', 'Immutable.Set', 'Immutable.List'],
+          newIsCap: true,
+          newIsCapExceptions: [],
+        },
+      ],
+      'no-alert': 'warn',
+      'no-array-constructor': 'error',
+      'no-await-in-loop': 'error',
+      'no-bitwise': 'error',
+      'no-caller': 'error',
+      'no-cond-assign': ['error', 'always'],
+      'no-console': 'warn',
+      'no-constant-condition': 'warn',
+      'no-continue': 'error',
+      'no-duplicate-imports': 'error',
+      'no-else-return': ['error', { allowElseIf: true }],
+      'no-empty-function': ['error', { allow: ['arrowFunctions', 'functions', 'methods'] }],
+      'no-eval': 'error',
+      'no-extend-native': 'error',
+      'no-extra-bind': 'error',
+      'no-extra-label': 'error',
+      'no-implied-eval': 'error',
+      'no-iterator': 'error',
+      'no-label-var': 'error',
+      'no-labels': ['error', { allowLoop: false, allowSwitch: false }],
+      'no-lone-blocks': 'error',
+      'no-lonely-if': 'error',
+      'no-loop-func': 'error',
+      'no-multi-assign': ['error'],
+      'no-multi-str': 'error',
+      'no-nested-ternary': 'error',
+      'no-new': 'error',
+      'no-new-func': 'error',
+      'no-new-wrappers': 'error',
+      'no-object-constructor': 'error',
+      'no-octal-escape': 'error',
+      'no-param-reassign': ['error', { props: false }],
+      'no-plusplus': 'error',
+      'no-proto': 'error',
+      'no-prototype-builtins': 'off',
+      'no-restricted-globals': [
+        'error',
+        'isFinite',
+        'isNaN',
+        'addEventListener',
+        'blur',
+        'close',
+        'closed',
+        'confirm',
+        'defaultStatus',
+        'event',
+        'external',
+        'defaultstatus',
+        'find',
+        'focus',
+        'frameElement',
+        'frames',
+        'history',
+        'innerHeight',
+        'innerWidth',
+        'length',
+        'location',
+        'locationbar',
+        'menubar',
+        'moveBy',
+        'moveTo',
+        'name',
+        'onblur',
+        'onerror',
+        'onfocus',
+        'onload',
+        'onresize',
+        'onunload',
+        'open',
+        'opener',
+        'opera',
+        'outerHeight',
+        'outerWidth',
+        'pageXOffset',
+        'pageYOffset',
+        'parent',
+        'print',
+        'removeEventListener',
+        'resizeBy',
+        'resizeTo',
+        'screen',
+        'screenLeft',
+        'screenTop',
+        'screenX',
+        'screenY',
+        'scroll',
+        'scrollbars',
+        'scrollBy',
+        'scrollTo',
+        'scrollX',
+        'scrollY',
+        'self',
+        'status',
+        'statusbar',
+        'stop',
+        'toolbar',
+        'top',
+      ],
+      'no-restricted-imports': ['error'],
+      'no-restricted-properties': [
+        'error',
+        {
+          message: 'arguments.callee is deprecated',
+          object: 'arguments',
+          property: 'callee',
+        },
+        {
+          message: 'Please use Number.isFinite instead',
+          object: 'global',
+          property: 'isFinite',
+        },
+        {
+          message: 'Please use Number.isFinite instead',
+          object: 'self',
+          property: 'isFinite',
+        },
+        {
+          message: 'Please use Number.isFinite instead',
+          object: 'window',
+          property: 'isFinite',
+        },
+        {
+          message: 'Please use Number.isNaN instead',
+          object: 'global',
+          property: 'isNaN',
+        },
+        {
+          message: 'Please use Number.isNaN instead',
+          object: 'self',
+          property: 'isNaN',
+        },
+        {
+          message: 'Please use Number.isNaN instead',
+          object: 'window',
+          property: 'isNaN',
+        },
+        {
+          message: 'Please use Object.defineProperty instead.',
+          property: '__defineGetter__',
+        },
+        {
+          message: 'Please use Object.defineProperty instead.',
+          property: '__defineSetter__',
+        },
+        {
+          message: 'Use the exponentiation operator (**) instead.',
+          object: 'Math',
+          property: 'pow',
+        },
+      ],
+      'no-restricted-syntax': [
+        'error',
+        {
+          message:
+            'for..in loops iterate over the entire prototype chain, which is virtually never what you want. Use Object.{keys,values,entries}, and iterate over the resulting array.',
+          selector: 'ForInStatement',
+        },
+        {
+          message:
+            'iterators/generators require regenerator-runtime, which is too heavyweight for this guide to allow them. Separately, loops should be avoided in favor of array iterations.',
+          selector: 'ForOfStatement',
+        },
+        {
+          message:
+            'Labels are a form of GOTO; using them makes code confusing and hard to maintain and understand.',
+          selector: 'LabeledStatement',
+        },
+        {
+          message:
+            '`with` is disallowed in strict mode because it makes code impossible to predict and optimize.',
+          selector: 'WithStatement',
+        },
+        {
+          message: 'Property setters are not allowed',
+          selector: 'MethodDefinition[kind="set"]',
+        },
+        {
+          message: 'Property getters are not allowed',
+          selector: 'MethodDefinition[kind="get"]',
+        },
+      ],
+      'no-return-assign': ['error', 'except-parens'],
+      'no-script-url': 'error',
+      'no-self-compare': 'error',
+      'no-sequences': 'error',
+      'no-shadow': ['error', { hoist: 'all' }],
+      'no-template-curly-in-string': 'error',
+      'no-undef-init': 'error',
+      'no-underscore-dangle': ['error', { allowAfterSuper: true, allowAfterThis: true }],
+      'no-unneeded-ternary': ['error', { defaultAssignment: false }],
+      'no-unused-expressions': 'error',
+      'no-unused-vars': ['error', { args: 'after-used', ignoreRestSiblings: true, vars: 'all' }],
+      'no-use-before-define': ['error', 'nofunc'],
+      'no-useless-computed-key': 'error',
+      'no-useless-concat': 'error',
+      'no-useless-constructor': 'error',
+      'no-useless-rename': [
+        'error',
+        { ignoreDestructuring: false, ignoreExport: false, ignoreImport: false },
+      ],
+      'no-useless-return': 'error',
+      'no-var': 'error',
+      'no-void': ['error', { allowAsStatement: true }],
+      'object-shorthand': ['error', 'always', { avoidQuotes: true, ignoreConstructors: false }],
+      'one-var': ['error', 'never'],
+      'operator-assignment': ['error', 'always'],
+      'prefer-arrow-callback': ['error', { allowNamedFunctions: false, allowUnboundThis: true }],
+      'prefer-const': ['error', { destructuring: 'any', ignoreReadBeforeAssign: true }],
+      'prefer-numeric-literals': 'error',
+      'prefer-rest-params': 'error',
+      'prefer-template': 'error',
+      radix: 'error',
+      'sort-imports': ['error', { ignoreCase: true, ignoreDeclarationSort: true }],
+      strict: 'error',
+      'switch-case/newline-between-switch-case': ['warn', 'always', { fallthrough: 'never' }],
+      'symbol-description': 'error',
+      'valid-jsdoc': ['error', { requireParamDescription: false, requireReturnDescription: false }],
+      'valid-typeof': ['error', { requireStringLiterals: true }],
+      'vars-on-top': 'error',
+      yoda: 'error',
+    },
   },
-  extends: ['eslint:recommended', 'plugin:import/errors', 'plugin:import/warnings'],
-  parserOptions: {
-    ecmaVersion: 'latest',
-    sourceType: 'module',
-  },
-  plugins: ['import', 'gettext', 'switch-case', '@stylistic'],
-  reportUnusedDisableDirectives: true,
-  rules: {
-    '@stylistic/padding-line-between-statements': [
-      'warn',
-      {
-        blankLine: 'always',
-        next: 'return',
-        prev: '*',
-      },
-      {
-        blankLine: 'always',
-        next: ['block', 'block-like', 'class', 'const', 'let', 'var', 'interface'],
-        prev: '*',
-      },
-      {
-        blankLine: 'always',
-        next: '*',
-        prev: ['block', 'block-like', 'class', 'const', 'let', 'var', 'interface'],
-      },
-      {
-        blankLine: 'any',
-        next: ['const', 'let', 'var'],
-        prev: ['const', 'let', 'var'],
-      },
-      {
-        blankLine: 'any',
-        next: ['case'],
-        prev: ['case'],
-      },
-      {
-        blankLine: 'always',
-        next: '*',
-        prev: 'directive',
-      },
-      {
-        blankLine: 'any',
-        next: 'directive',
-        prev: 'directive',
-      },
-    ],
-    '@stylistic/spaced-comment': [
-      'warn',
-      'always',
-      {
-        block: { balanced: true, exceptions: ['-', '+'], markers: ['=', '!'] },
-        line: { exceptions: ['-', '+'], markers: ['=', '!'] },
-      },
-    ],
-    'array-callback-return': 'error',
-    'arrow-body-style': ['error', 'as-needed', { requireReturnForObjectLiteral: false }],
-    'block-scoped-var': 'error',
-    camelcase: ['error', { properties: 'never' }],
-    complexity: 'error',
-    'consistent-return': 'error',
-    'default-case': ['error'],
-    'default-param-last': 'error',
-    'dot-notation': ['error', { allowKeywords: true }],
-    eqeqeq: ['error', 'smart'],
-    'func-names': 'error',
-    'getter-return': ['error', { allowImplicit: true }],
-    'gettext/no-variable-string': 'error',
-    'guard-for-in': 'error',
-    'import/default': 'off',
-    'import/dynamic-import-chunkname': 'error',
-    'import/extensions': [
-      'error',
-      {
-        js: 'never',
-        json: 'always',
-        jsx: 'never',
-        ts: 'never',
-        tsx: 'never',
-      },
-    ],
-    'import/newline-after-import': 'warn',
-    'import/no-absolute-path': 'error',
-    'import/no-amd': 'error',
-    'import/no-duplicates': ['warn', { 'prefer-inline': true }],
-    'import/no-dynamic-require': 'error',
-    'import/no-extraneous-dependencies': 'error',
-    'import/no-mutable-exports': 'error',
-    'import/no-named-default': 'error',
-    'import/no-unresolved': ['error', { caseSensitive: true, commonjs: true }],
-    'import/no-webpack-loader-syntax': 'error',
-    'import/order': [
-      'warn',
-      {
-        alphabetize: { caseInsensitive: true, order: 'asc' },
-        groups: [['builtin', 'external'], 'internal', 'parent', 'sibling', 'index'],
-        'newlines-between': 'always',
-      },
-    ],
-    'max-classes-per-file': ['error', 1],
-    'new-cap': [
-      'error',
-      {
-        capIsNew: false,
-        capIsNewExceptions: ['Immutable.Map', 'Immutable.Set', 'Immutable.List'],
-        newIsCap: true,
-        newIsCapExceptions: [],
-      },
-    ],
-    'no-alert': 'warn',
-    'no-array-constructor': 'error',
-    'no-await-in-loop': 'error',
-    'no-bitwise': 'error',
-    'no-caller': 'error',
-    'no-cond-assign': ['error', 'always'],
-    'no-console': 'warn',
-    'no-constant-condition': 'warn',
-    'no-continue': 'error',
-    'no-duplicate-imports': 'error',
-    'no-else-return': ['error', { allowElseIf: true }],
-    'no-empty-function': ['error', { allow: ['arrowFunctions', 'functions', 'methods'] }],
-    'no-eval': 'error',
-    'no-extend-native': 'error',
-    'no-extra-bind': 'error',
-    'no-extra-label': 'error',
-    'no-implied-eval': 'error',
-    'no-iterator': 'error',
-    'no-label-var': 'error',
-    'no-labels': ['error', { allowLoop: false, allowSwitch: false }],
-    'no-lone-blocks': 'error',
-    'no-lonely-if': 'error',
-    'no-loop-func': 'error',
-    'no-multi-assign': ['error'],
-    'no-multi-str': 'error',
-    'no-nested-ternary': 'error',
-    'no-new': 'error',
-    'no-new-func': 'error',
-    'no-new-wrappers': 'error',
-    'no-object-constructor': 'error',
-    'no-octal-escape': 'error',
-    'no-param-reassign': ['error', { props: false }],
-    'no-plusplus': 'error',
-    'no-proto': 'error',
-    'no-prototype-builtins': 'off',
-    'no-restricted-globals': [
-      'error',
-      'isFinite',
-      'isNaN',
-      'addEventListener',
-      'blur',
-      'close',
-      'closed',
-      'confirm',
-      'defaultStatus',
-      'event',
-      'external',
-      'defaultstatus',
-      'find',
-      'focus',
-      'frameElement',
-      'frames',
-      'history',
-      'innerHeight',
-      'innerWidth',
-      'length',
-      'location',
-      'locationbar',
-      'menubar',
-      'moveBy',
-      'moveTo',
-      'name',
-      'onblur',
-      'onerror',
-      'onfocus',
-      'onload',
-      'onresize',
-      'onunload',
-      'open',
-      'opener',
-      'opera',
-      'outerHeight',
-      'outerWidth',
-      'pageXOffset',
-      'pageYOffset',
-      'parent',
-      'print',
-      'removeEventListener',
-      'resizeBy',
-      'resizeTo',
-      'screen',
-      'screenLeft',
-      'screenTop',
-      'screenX',
-      'screenY',
-      'scroll',
-      'scrollbars',
-      'scrollBy',
-      'scrollTo',
-      'scrollX',
-      'scrollY',
-      'self',
-      'status',
-      'statusbar',
-      'stop',
-      'toolbar',
-      'top',
-    ],
-    'no-restricted-imports': ['error'],
-    'no-restricted-properties': [
-      'error',
-      {
-        message: 'arguments.callee is deprecated',
-        object: 'arguments',
-        property: 'callee',
-      },
-      {
-        message: 'Please use Number.isFinite instead',
-        object: 'global',
-        property: 'isFinite',
-      },
-      {
-        message: 'Please use Number.isFinite instead',
-        object: 'self',
-        property: 'isFinite',
-      },
-      {
-        message: 'Please use Number.isFinite instead',
-        object: 'window',
-        property: 'isFinite',
-      },
-      {
-        message: 'Please use Number.isNaN instead',
-        object: 'global',
-        property: 'isNaN',
-      },
-      {
-        message: 'Please use Number.isNaN instead',
-        object: 'self',
-        property: 'isNaN',
-      },
-      {
-        message: 'Please use Number.isNaN instead',
-        object: 'window',
-        property: 'isNaN',
-      },
-      {
-        message: 'Please use Object.defineProperty instead.',
-        property: '__defineGetter__',
-      },
-      {
-        message: 'Please use Object.defineProperty instead.',
-        property: '__defineSetter__',
-      },
-      {
-        message: 'Use the exponentiation operator (**) instead.',
-        object: 'Math',
-        property: 'pow',
-      },
-    ],
-    'no-restricted-syntax': [
-      'error',
-      {
-        message:
-          'for..in loops iterate over the entire prototype chain, which is virtually never what you want. Use Object.{keys,values,entries}, and iterate over the resulting array.',
-        selector: 'ForInStatement',
-      },
-      {
-        message:
-          'iterators/generators require regenerator-runtime, which is too heavyweight for this guide to allow them. Separately, loops should be avoided in favor of array iterations.',
-        selector: 'ForOfStatement',
-      },
-      {
-        message:
-          'Labels are a form of GOTO; using them makes code confusing and hard to maintain and understand.',
-        selector: 'LabeledStatement',
-      },
-      {
-        message:
-          '`with` is disallowed in strict mode because it makes code impossible to predict and optimize.',
-        selector: 'WithStatement',
-      },
-      {
-        message: 'Property setters are not allowed',
-        selector: 'MethodDefinition[kind="set"]',
-      },
-      {
-        message: 'Property getters are not allowed',
-        selector: 'MethodDefinition[kind="get"]',
-      },
-    ],
-    'no-return-assign': ['error', 'except-parens'],
-    'no-script-url': 'error',
-    'no-self-compare': 'error',
-    'no-sequences': 'error',
-    'no-shadow': ['error', { hoist: 'all' }],
-    'no-template-curly-in-string': 'error',
-    'no-undef-init': 'error',
-    'no-underscore-dangle': ['error', { allowAfterSuper: true, allowAfterThis: true }],
-    'no-unneeded-ternary': ['error', { defaultAssignment: false }],
-    'no-unused-expressions': 'error',
-    'no-unused-vars': ['error', { args: 'after-used', ignoreRestSiblings: true, vars: 'all' }],
-    'no-use-before-define': ['error', 'nofunc'],
-    'no-useless-computed-key': 'error',
-    'no-useless-concat': 'error',
-    'no-useless-constructor': 'error',
-    'no-useless-rename': [
-      'error',
-      { ignoreDestructuring: false, ignoreExport: false, ignoreImport: false },
-    ],
-    'no-useless-return': 'error',
-    'no-var': 'error',
-    'no-void': ['error', { allowAsStatement: true }],
-    'object-shorthand': ['error', 'always', { avoidQuotes: true, ignoreConstructors: false }],
-    'one-var': ['error', 'never'],
-    'operator-assignment': ['error', 'always'],
-    'prefer-arrow-callback': ['error', { allowNamedFunctions: false, allowUnboundThis: true }],
-    'prefer-const': ['error', { destructuring: 'any', ignoreReadBeforeAssign: true }],
-    'prefer-numeric-literals': 'error',
-    'prefer-rest-params': 'error',
-    'prefer-template': 'error',
-    radix: 'error',
-    'sort-imports': ['error', { ignoreCase: true, ignoreDeclarationSort: true }],
-    strict: 'error',
-    'switch-case/newline-between-switch-case': ['warn', 'always', { fallthrough: 'never' }],
-    'symbol-description': 'error',
-    'valid-jsdoc': ['error', { requireParamDescription: false, requireReturnDescription: false }],
-    'valid-typeof': ['error', { requireStringLiterals: true }],
-    'vars-on-top': 'error',
-    yoda: 'error',
-  },
-};
+];
