@@ -1,5 +1,13 @@
-/** @type {import("eslint").Linter.Config} */
-module.exports = {
-  extends: 'plugin:jest-dom/recommended',
-  plugins: ['jest-dom'],
-};
+import jestDomPlugin from 'eslint-plugin-jest-dom';
+
+export default [
+  {
+    files: ['**/*.spec.*', '**/spec.*', 'jest-setup.*'],
+    plugins: {
+      'jest-dom': jestDomPlugin,
+    },
+    rules: {
+      ...jestDomPlugin.configs.recommended.rules,
+    },
+  },
+];
