@@ -1,8 +1,14 @@
-/** @type {import("eslint").Linter.Config} */
-module.exports = {
-  extends: ['plugin:testing-library/react'],
-  plugins: ['testing-library'],
-  rules: {
-    'testing-library/no-render-in-lifecycle': 'off',
+import testingLibraryPlugin from 'eslint-plugin-testing-library';
+
+export default [
+  {
+    files: ['**/*.spec.*', '**/spec.*', 'jest-setup.*'],
+    plugins: {
+      'testing-library': testingLibraryPlugin,
+    },
+    rules: {
+      ...testingLibraryPlugin.configs.react.rules,
+      'testing-library/no-render-in-lifecycle': 'off',
+    },
   },
-};
+];
