@@ -1,6 +1,8 @@
 /** @type {import('jest').Config} */
 const config = {
   testPathIgnorePatterns: ['/__.*/', '<rootDir>/node_modules/'],
+  transform: {},
+  testEnvironment: 'node',
 };
 
-module.exports = config;
+export default config;
