@@ -1,13 +1,11 @@
-function hasPackage(pkgName) {
+export async function hasPackage(pkgName) {
   try {
-    require.resolve(pkgName, { paths: [process.cwd()] });
+    // In ES modules, we try to resolve the package from the current working directory
+    // Note: import.meta.resolve is experimental, so we'll use dynamic import instead
+    await import(pkgName);
 
     return true;
-  } catch (e) {
+  } catch {
     return false;
   }
 }
-
-module.exports = {
-  hasPackage,
-};
